Adding support for creating new games on the server and joining existing games.  Debating whether Games should be serialized and what they actually represent.  Perhaps players don't join Games, they join GameInstances.  I don't know.  Must talk to Bryan.
MenuState and ServerConnection are bloating.

Lots of minor type-safety changes.


git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@212 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/base/Socket.cpp
@@ -2,7 +2,6 @@
 #include <prnetdb.h>
 #include "NSPRUtility.h"
 #include "Socket.h"
-#include "Types.h"
 
 
 namespace pyr {
@@ -11,7 +10,7 @@ namespace pyr {
         _socket = fd;
     }
     
-    Socket::Socket(const std::string& host, int port) {
+    Socket::Socket(const std::string& host, u16 port) {
         char buffer[PR_NETDB_BUF_SIZE];
         PRHostEnt hostentry;
         PRStatus status = PR_GetHostByName(
